refactor: header drawer imports & index export
