+++ src/components/UI/molecules/HeaderDrawerCompany/index.ts
@@ -1,1 +1,2 @@
-export * from './HeaderDrawerCompany.components'
+export * from './HeaderDrawerCompany.interface'
+export { HeaderDrawerCompany } from './HeaderDrawerCompany.components'
